+++ src/snowflake/cli/plugins/snowpark/package/commands.py
@@ -6,6 +6,7 @@ from pathlib import Path
 import typer
 from snowflake.cli.api.commands.snow_typer import SnowTyper
 from snowflake.cli.api.output.types import CommandResult, MessageResult
+from snowflake.cli.plugins.snowpark.models import PypiOption
 from snowflake.cli.plugins.snowpark.package.manager import (
     cleanup_after_install,
     create,
@@ -17,6 +18,7 @@ from snowflake.cli.plugins.snowpark.package.utils import (
     NotInAnaconda,
     RequiresPackages,
 )
+from snowflake.cli.plugins.snowpark.snowpark_shared import PackageNativeLibrariesOption
 
 app = SnowTyper(
     name="package",
@@ -37,6 +39,7 @@ install_option = typer.Option(
 def package_lookup(
     name: str = typer.Argument(..., help="Name of the package."),
     install_packages: bool = install_option,
+    allow_native_libraries: PypiOption = PackageNativeLibrariesOption,
     **options,
 ) -> CommandResult:
     """
@@ -44,7 +47,11 @@ def package_lookup(
     If the `--yes` flag is provided, this command checks all dependencies of the packages
     outside Snowflake channel.
     """
-    lookup_result = lookup(name=name, install_packages=install_packages)
+    lookup_result = lookup(
+        name=name,
+        install_packages=install_packages,
+        allow_native_libraries=allow_native_libraries,
+    )
     cleanup_after_install()
     return MessageResult(lookup_result.message)
 
@@ -85,6 +92,7 @@ def package_create(
         help="Name of the package to create.",
     ),
     install_packages: bool = install_option,
+    allow_native_libraries: PypiOption = PackageNativeLibrariesOption,
     **options,
 ) -> CommandResult:
     """
@@ -92,7 +100,13 @@ def package_create(
     """
 
     if (
-        type(lookup_result := lookup(name=name, install_packages=install_packages))
+        type(
+            lookup_result := lookup(
+                name=name,
+                install_packages=install_packages,
+                allow_native_libraries=allow_native_libraries,
+            )
+        )
         in [
             NotInAnaconda,
             RequiresPackages,